Fixed response codes for independent calculation servlet. Will now show code 409 (conflict) if there was an error.
Changed Operation names from uppercase to capitalized to fit assignment description.
Added equalsIgnoreCase to comparison between operation name and attempted operation.
An error that was encountered when an operation was matched will now show the matched operation name in the error message instead of the attempted operation name (to avoid showing "pLus" instead of "Plus")